move customeColors.ts and update lint commands
