+++ app/benchmarking/components/benchmarkReportViewer.tsx
@@ -4,9 +4,9 @@ import { downloadHtmlReport } from '@/app/benchmarking/utils/reportDownloader';
 import { Button, ButtonType } from '@/app/components/button';
 import { MainSectionSurface } from '@/app/components/mainSectionSurface';
 import { SelectInput } from '@/app/components/selectInput';
+import { colors } from '@/app/customColors';
 import { useGetBenchmarksResultQuery } from '@/app/services/benchmark-api-service';
 import { useGetRunnerByIdQuery } from '@/app/services/runner-api-service';
-import { colors } from '@/app/views/shared-components/customColors';
 import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
 import { BenchmarkReport } from './benchmarkReport';
