+++ app/benchmarking/components/benchmarkRunsView.tsx
@@ -6,8 +6,8 @@ import React, { CSSProperties } from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
 import { Button, ButtonType } from '@/app/components/button';
 import { MainSectionSurface } from '@/app/components/mainSectionSurface';
+import { colors } from '@/app/customColors';
 import { formatDateFromTimestamp } from '@/app/lib/date-utils';
-import { colors } from '@/app/views/shared-components/customColors';
 
 interface CustomStyle extends CSSProperties {
   WebkitLineClamp?: string;
