+++ app/benchmarking/components/cookbookSelectionItem.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
-import { colors } from '@/app/views/shared-components/customColors';
+import { colors } from '@/app/customColors';
 
 type CookbookSelectionItemProps = {
   cookbook: Cookbook;
