+++ app/benchmarking/recipes/recipesViewList.tsx
@@ -8,7 +8,7 @@ import { Icon, IconName } from '@/app/components/IconSVG';
 import { Button, ButtonType } from '@/app/components/button';
 import { Modal } from '@/app/components/modal';
 import { TextInput } from '@/app/components/textInput';
-import { colors } from '@/app/views/shared-components/customColors';
+import { colors } from '@/app/customColors';
 import { Step } from './enums';
 import { SelectedRecipesPills } from './selectedRecipesPills';
