+++ app/history/error.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
 import { Modal } from '@/app/components/modal';
-import { colors } from '@/app/views/shared-components/customColors';
+import { colors } from '@/app/customColors';
 
 export default function Error({
   error,
