+++ app/redteaming/(fullscreen)/components/prompt-box.tsx
@@ -3,10 +3,10 @@ import { Button, ButtonType } from '@/app/components/button';
 import { ListItem, SelectList } from '@/app/components/selectList';
 import { TextArea } from '@/app/components/textArea';
 import { Tooltip, TooltipPosition } from '@/app/components/tooltip';
+import { colors } from '@/app/customColors';
 import useOutsideClick from '@/app/hooks/use-outside-click';
 import { debounce } from '@/app/lib/throttle';
 import useChatboxesPositionsUtils from '@/app/redteaming/(fullscreen)/hooks/useChatboxesPositionsUtils';
-import { colors } from '@/app/views/shared-components/customColors';
 import { toggleDarkMode, useAppDispatch } from '@/lib/redux';
 import {
   LayoutMode,
