+++ app/redteaming/(fullscreen)/components/promptTemplatesList.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
 import { Button, ButtonType } from '@/app/components/button';
+import { colors } from '@/app/customColors';
 import { useGetAllPromptTemplatesQuery } from '@/app/services/prompt-template-api-service';
-import { colors } from '@/app/views/shared-components/customColors';
 import { LoadingAnimation } from '@/app/views/shared-components/loadingAnimation';
 
 type PromptTemplatesListProps = {
