+++ app/redteaming/(fullscreen)/components/saveBookmarkModal.tsx
@@ -6,7 +6,7 @@ import { Button, ButtonType } from '@/app/components/button';
 import { Chat } from '@/app/components/chat';
 import { Modal } from '@/app/components/modal';
 import { TextInput } from '@/app/components/textInput';
-import { colors } from '@/app/views/shared-components/customColors';
+import { colors } from '@/app/customColors';
 import { PromptBubbleInfo } from './prompt-bubble-info';
 import { SubmitButton } from './submitButton';
