+++ app/redteaming/(others)/sessions/new/redteamRunForm.tsx
@@ -5,8 +5,8 @@ import { object, string, array } from 'yup';
 import { Button, ButtonType } from '@/app/components/button';
 import { TextArea } from '@/app/components/textArea';
 import { TextInput } from '@/app/components/textInput';
+import { colors } from '@/app/customColors';
 import { useCreateSessionMutation } from '@/app/services/session-api-service';
-import { colors } from '@/app/views/shared-components/customColors';
 import {
   resetAttackModule,
   resetRedteamModels,
