+++ app/redteaming/(others)/sessions/redteamSessionsViewList.tsx
@@ -4,7 +4,7 @@ import React, { CSSProperties, useState } from 'react';
 import { Icon, IconName } from '@/app/components/IconSVG';
 import { Button, ButtonType } from '@/app/components/button';
 import { MainSectionSurface } from '@/app/components/mainSectionSurface';
-import { colors } from '@/app/views/shared-components/customColors';
+import { colors } from '@/app/customColors';
 
 interface CustomStyle extends CSSProperties {
   webkitLineClamp?: string;
