+++ app/utilities/page.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { IconName } from '@/app/components/IconSVG';
 import BackToHomeButton from '@/app/components/backToHomeButton';
 import { SubmenuButton } from '@/app/components/submenuButton/submenuButton';
-import { colors } from '@/app/views/shared-components/customColors';
+import { colors } from '@/app/customColors';
 
 export default function UtilitiesHomePage() {
   return (
