+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "start-trace": "node --trace-warnings node_modules/.bin/next",
-    "lint": "next lint --quiet && npm run tsc",
+    "lint": "next lint && npm run tsc",
     "lint:fix": "next lint --fix",
     "tsc": "tsc --noEmit",
     "test": "jest",
